Lots of improvements/pathces.
Imported git aliases from simon-manjaroawesome.

Megascript is no longer one megascript, but 2. The reason is that when I tried to use this to setup a fresh install of Ubuntu WSL Edition, nothing happened after installing omz.

I made a Powershell script that will help setting up WSL, as well as
instructions for how to upgrade to WSL2. Why Microsoft don't let you
just install it as a feature like with WSL(1) is beyond me.

Another new script is the "ssh-setup.sh" script. It will make the SSH
service start and will start automatically from there on, then help
setting up your github account.

Besides that, I've added a setup for tabnine (an AI developer tool for
different IDEs and editors that will give you suggestions as you type
your text or code) for vim and for the colorscheme badwolf.

Also added tldr to the install script. A tool for when you don't feel
like reading through a whole manpage.